% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicFun.R
\name{bayesTops.predict}
\alias{bayesTops.predict}
\title{Predict labels using a trained BayesToPs model.}
\usage{
bayesTops.predict(trained.model, newdata)
}
\arguments{
\item{trained.model}{Trained model}

\item{mewdata}{New dataset with the same features}
}
\value{
list with raw and normalised predicted labels
}
\description{
Predict labels using a trained BayesToPs model.
}
