% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicFun.R
\name{bayesTops}
\alias{bayesTops}
\title{Train a BayesToPs model for regression.}
\usage{
bayesTops(x, y, initialPrior, normalise = TRUE,
  features2norm = names(x), normaliseY = TRUE, decay = 0.1,
  blrResVariance = 1, modelSelection = "Bayes factors",
  propagatePosterior = TRUE, weightsMethod = "mean",
  metric_createTree = "rmse", metric_finalWeights = "rmse",
  maxDepthTree = 10, maxSizeSlices = 500, maxSplitValueTested = 50,
  minSizeLeaf = 1, minNbObsV2 = 1, printProgress = TRUE)
}
\arguments{
\item{x}{A DataFrame with n rows and p columns}

\item{y}{A vector of length n with the labels}

\item{initialPrior}{A list with an attribute mean (vector of length p)
and an attribute variance (square matrix of dimension p).
It's the initial prior to be used on the root of the tree.}

\item{normalise}{a boolean, if TRUE (default) then the data is normalised
beforehand.}

\item{features2norm}{vector containing the columns to normalise.
Ignored if normalise=FALSE.}

\item{normaliseY}{a boolean, if TRUE (default), then the response
variable is also normalised. Ignored if normalise=FALSE.}

\item{decay}{a float between 0 and 1 fixing the decay parameter.}

\item{blrResVariance}{a float representing the residual variance in the
Bayesian Linear Model fitted at each node.}

\item{modelSelection}{one of "Bayes factors" or "Validation",
describe the model selection method used to divide the nodes.}

\item{propagatePosterior}{a boolean, if TRUE (default),
the posterior distribution is propagated to the child nodes,
otherwise initialPrior is used as prior.}

\item{weightsMethod}{one of "lm", "validation", "bayes",
"leaf only" or "mean" (default). Decided how to find the final weight of
the predictor in each leaf.}

\item{metric_createTree}{one of "auc", "rmse" (default), "r2" or "mae".
The metric used to evaluate the performance of the models
when building the tree.}

\item{metric_finalWeights}{one of "auc", "rmse" (default), "r2" or "mae".
The metric used to evaluate the performance of the models
when fitting the final weights.}

\item{maxDepthTree}{int limiting the depth of the tree (default to 10).}

\item{maxSizeSlices}{int limiting the maximum number of observations to
consider when computing the joint marginal likelihood at a node
(default to 500).}

\item{maxSplitValueTested}{int limiting the number of split value
to test at each node (default to 50).}

\item{minSizeLeaf}{int fixing the minimum number of observations
from the training set in a leaf (default to 1).}

\item{minNbObsV2}{int, only if weightsMethod="validation".}

\item{printProgress}{boolean, if TRUE (default), the progress of the
tree will be displayed.}
}
\value{
A trained BayesToPs model
}
\description{
Train a BayesToPs model for regression.
}
