% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{servo}
\alias{servo}
\title{Data from the servo dataset}
\format{A data frame with .. rows and .. variables:
\describe{
  \item{v1}{blabla}
  \item{v2}{}
  ...
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Servo}
}
\usage{
servo
}
\description{
[description of servo]
}
\keyword{datasets}
